Some datastructures to detect which objects are live.
Live = part of the snapshot we want to create.

We use a bitmap that marks all bits of the live
objects, which will be useful later when handling
gaps etc.

It should be possible to mark most objects with a
single bitwise operation on one 64 bit word.  For now
it's an inefficient loop.

Code is untested, just to show direction.