Make framework that generates both standalone and mem-backed snapshots.
Try to refactor so we can generate both standalone snapshots
that contain the heap labs, and also memory-backed ones
where the original snapshot is used as backing for a relocating
memcpy or an mmap operation.

Reverts the abuse of heap space enum and introduces a new one
that is just for this.

During serialization: A forwarding map to find out where objects went.